Move listings jsons to data/ directory
