+++ services/listings/src/lib/ami_charts.ts
@@ -1,7 +1,7 @@
-import sanMateoHUD2019 from "../../ami_charts/SanMateoHUD2019.json"
-import sanMateoHOME2019 from "../../ami_charts/SanMateoHOME2019.json"
-import sanMateoHERASpecial2019 from "../../ami_charts/SanMateoHERASpecial2019.json"
-import sanJoseTCAC2019 from "../../ami_charts/SanJoseTCAC2019.json"
+import sanMateoHUD2019 from "../../data/ami_charts/SanMateoHUD2019.json"
+import sanMateoHOME2019 from "../../data/ami_charts/SanMateoHOME2019.json"
+import sanMateoHERASpecial2019 from "../../data/ami_charts/SanMateoHERASpecial2019.json"
+import sanJoseTCAC2019 from "../../data/ami_charts/SanJoseTCAC2019.json"
 
 export const amiCharts = {
   1: sanMateoHUD2019,
